add pthread

+++ src/test/vixMntOperation_test.cpp
@@ -18,8 +18,8 @@ main(int argc,char** argv){
 
         char buf[] = "ashione";
         VixMntOpRead t("/tmp",buf,3,2);
-        printf("OpRead struct size : %d %d %d\n",sizeof(VixMntOpRead),t.size(),sizeof(t));
-        printf("buf addr : %x %s\n",buf,buf);
+        printf("OpRead struct size : %ld %d %ld\n",sizeof(VixMntOpRead),t.size(),sizeof(t));
+        printf("buf addr : %x %s\n",( long )buf,buf);
         VixMntMsgQue* msgQue = new VixMntMsgQue("/op");
         VixMntMsgData* msgdata = new VixMntMsgData(VixMntMsgOp::MntInit,sizeof(t),(char *)&t);
         msgQue->sendMsg(msgdata);
